pyproject.toml
[project]
name = "putong-ai-overseer"
version = "3.0.0"
description = "AI algorithm problem evaluation tool with virtual judge"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "rich>=13.9.4",
    "openai>=1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["judger", "overseer"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
